Move data API test helpers to the test lib
