+++ tracdap-libs/tracdap-lib-test/build.gradle
@@ -34,6 +34,16 @@ dependencies {
     api project(':tracdap-lib-db')
     api project(':tracdap-lib-data')
 
+    // Tests need the ability to spin up instances of the TRAC services to test against
+    api project(':tracdap-svc-meta')
+    api project(':tracdap-svc-data')
+    api project(':tracdap-svc-orch')
+    api project(':tracdap-gateway')
+
+    // Give tests access to the deploy tools to set up services for integration
+    api project(':deploy-metadb')
+    api project(':secret-tool')
+
     implementation group: 'org.junit.jupiter', name: 'junit-jupiter-api', version: "$junit_version"
     implementation group: 'org.junit.jupiter', name: 'junit-jupiter-params', version: "$junit_version"
 
@@ -43,26 +53,19 @@ dependencies {
     implementation group: 'io.netty', name: 'netty-handler-proxy', version: "$netty_version"
     implementation group: 'io.grpc', name: 'grpc-netty', version: "$grpc_version"
 
-    // Use Flyway to deploy schemas in test cases
-    implementation group: 'org.flywaydb', name: 'flyway-core', version: "$flyway_version"
+    // Apache Arrow data framework
+    implementation group: 'org.apache.arrow', name: 'arrow-vector', version: "$arrow_version"
+    implementation group: 'org.apache.arrow', name: 'arrow-memory-netty', version: "$arrow_version"
 
     // Jackson uses runtime class resolution
     // Make sure classes needed for config parsing are always on the runtime classpath for tests
-    api group: 'com.fasterxml.jackson.core', name: 'jackson-core', version: "$jackson_version"
-    api group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: "$jackson_databind_version"
-    api group: 'com.fasterxml.jackson.dataformat', name: 'jackson-dataformat-yaml', version: "$jackson_version"
+    implementation group: 'com.fasterxml.jackson.core', name: 'jackson-core', version: "$jackson_version"
+    implementation group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: "$jackson_databind_version"
+    implementation group: 'com.fasterxml.jackson.dataformat', name: 'jackson-dataformat-yaml', version: "$jackson_version"
+    implementation group: 'com.fasterxml.jackson.dataformat', name: 'jackson-dataformat-csv', version: "$jackson_version"
 
-    // Tests need the ability to spin up instances of the TRAC services to test against
-
-    api project(':tracdap-svc-meta')
-    api project(':tracdap-svc-data')
-    api project(':tracdap-svc-orch')
-    api project(':tracdap-gateway')
-
-    // Give tests access to the deploy tools to set up services for integration
-
-    api project(':deploy-metadb')
-    api project(':secret-tool')
+    // Use Flyway to deploy schemas in test cases
+    implementation group: 'org.flywaydb', name: 'flyway-core', version: "$flyway_version"
 }
