+++ tracdap-services/tracdap-svc-data/src/test/java/org/finos/tracdap/svc/data/api/DataRoundTripTest.java
@@ -24,6 +24,7 @@ import org.finos.tracdap.common.config.ConfigManager;
 import org.finos.tracdap.common.plugin.PluginManager;
 import org.finos.tracdap.common.util.ResourceHelpers;
 import org.finos.tracdap.metadata.*;
+import org.finos.tracdap.test.data.DataApiTestHelpers;
 import org.finos.tracdap.test.data.SampleData;
 import org.finos.tracdap.test.helpers.StorageTestHelpers;
 import org.finos.tracdap.test.helpers.PlatformTest;
